+++ src/node_parser.cpp
@@ -6,7 +6,6 @@
 
 #include <_node_parser.h>
 
-#include <fstream>
 #include <iostream>
 
 namespace cfg::parser
@@ -16,21 +15,20 @@ namespace cfg::parser
     std::cout << "+ Reading nodes" << std::endl;
     const cfg::reader::SectionReader node_reader("Nodes", mesh_stream);
 
-    const auto node_header = parse_node_header(node_reader, mesh_stream, mode);
-    const auto nodes       = parse_node_blocks(node_reader, node_header, mesh_stream, mode, parallel);
-    validate_nodes(nodes, node_header, parallel);
+    // Read the nodes
+    const auto reader = make_node_reader(parallel);
+    const auto nodes  = reader(node_reader, mesh_stream, mode);
 
-    // Check that we read the Nodes section correctly -> we should read
-    // "$EndNodes"
+    // Check that we read the Nodes section correctly -> we should read "$EndNodes"
     std::string line;
     node_reader(mesh_stream) >> line;
     if (line != "$EndNodes")
     {
+      std::cout << line << std::endl;
       throw std::runtime_error("The Nodes section was read incorrectly");
     }
 
     // Report how many nodes we read
-    std::cout << "++ Rank " << parallel.rank << " read " << nodes.size() << " out of " << node_header.n_nodes
-              << " nodes" << std::endl;
+    std::cout << "++ Rank " << parallel.rank << " read " << nodes.size() << " nodes" << std::endl;
   }
 }  // namespace cfg::parser